requirements.txt
python-telegram-bot==22.0
qrcode[pil]==8.0
pydantic-settings==2.8.1
pydantic[email]==2.10.6